requirement.txt
Package             Version
------------------- -------
asgiref             3.10.0
Django              5.2.7
djangorestframework 3.16.1
pip                 24.0
sqlparse            0.5.3
